reply 수정 삭제

+++ src/main/java/kosta/mvc/service/ReplyServiceImpl.java
@@ -5,6 +5,8 @@ import java.util.List;
 
 import kosta.mvc.dao.ReplyDAO;
 import kosta.mvc.dao.ReplyDAOImpl;
+import kosta.mvc.dao.StudyDAO;
+import kosta.mvc.dao.StudyDAOImpl;
 import kosta.mvc.dto.PostReply;
 
 public class ReplyServiceImpl implements ReplyService {
@@ -20,6 +22,11 @@ public class ReplyServiceImpl implements ReplyService {
 
 	@Override
 	public void updateReply(PostReply reply) throws SQLException {
+		String dbId = replyDao.getDBUserId(reply.getpReplyNo());
+		if(!dbId.equals(reply.getUserId())) {
+			throw new SQLException("본인이 작성한 댓글만 수정할 수 있습니다.");
+		}
+		
 		if(replyDao.updateReply(reply) == 0) {
 			throw new SQLException("댓글이 수정되지 않았습니다");
 		}
@@ -27,7 +34,12 @@ public class ReplyServiceImpl implements ReplyService {
 	}
 
 	@Override
-	public void deleteReply(int replyNo) throws SQLException {
+	public void deleteReply(int replyNo, String userId) throws SQLException {
+		String dbId = replyDao.getDBUserId(replyNo);
+		if(!dbId.equals(userId)) {
+			throw new SQLException("본인이 작성한 댓글만 수정할 수 있습니다.");
+		}
+		
 		if(replyDao.deleteReply(replyNo) == 0) {
 			throw new SQLException("댓글이 삭제되지 않았습니다");
 		}
